adding first github action

+++ app/routers/vote.py
@@ -32,9 +32,9 @@ async def get_vote(vote: schemas.Vote, db: Session = Depends(get_db),
         db.commit()
         return {"message": "successfully added vote"}
     else:
-        # if not found_vote:
-        #     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-        #                         detail="vote does not exist")
+        if not found_vote:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                                detail="vote does not exist")
 
         vote_query.delete(synchronize_session=False)
         db.commit()
